Moved on-the-fly synchronization to Sync cog
