+++ Dockerfile
@@ -3,6 +3,7 @@ FROM python:3.7-alpine
 ENV PYTHONDONTWRITEBYTECODE 1
 ENV PYTHONUNBUFFERED 1
 RUN apk update && apk add gcc git postgresql-dev musl-dev tzdata
+RUN apk add jpeg-dev zlib-dev freetype-dev lcms2-dev openjpeg-dev tiff-dev tk-dev tcl-dev
 RUN apk add docker
 ENV TZ=Europe/Prague
